tests: Align with original repo
Apparently upstream does not run all the tests, the beforeAll() is not
even executed. On our side it was, causing a lot of errors related to
timeouts.
